% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEverse-pkg.R
\docType{package}
\name{iSEEverse-pkg}
\alias{iSEEverse}
\alias{iSEEverse-package}
\alias{iSEEverse-pkg}
\title{iSEEverse}
\description{
Meta package that installs (and loads) all iSEE-related packages.
}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Kevin Rue-Albrecht \email{kevinrue67@gmail.com} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})
  \item Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
}

}
\keyword{internal}
