% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEverse.R
\name{iSEEverse_logo}
\alias{iSEEverse_logo}
\title{The iSEEverse logo, using ASCII characters}
\usage{
iSEEverse_logo()
}
\value{
The character vector to be parsed in the startup message
}
\description{
The iSEEverse logo, using ASCII characters
}
\details{
Generated via https://patorjk.com/software/taag/#p=display&f=Big&t=iSEE
}
\examples{
iSEEverse_logo()
}
