% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEverse.R
\name{iSEEverse_update}
\alias{iSEEverse_update}
\title{Update iSEE and the whole iSEEverse}
\usage{
iSEEverse_update(repos = getOption("repos"))
}
\arguments{
\item{repos}{The repositories to use to check for updates.
Defaults to getOption("repos").}
}
\value{
An invisible NULL or the suggested commands to update the
corresponding packages.
}
\description{
Update iSEE and the whole iSEEverse
}
\examples{
\dontrun{
iSEEverse_update()
}
}
